test: spring bean 이동
